Initial set of changes on rpcplus to add streaming RPC support. The code was forked because the changes are minimal and seem right, and because of the backward compatibility.
The protocal is backward compatible on the wire (old clients and servers
will still work with this new library, provided they don't use
the streaming RPC feature).

On the server side, a new interface type is recognized:
func (t *T) method(args T1, sendNonFinalReply func(reply interface{}) error, finalReply *T2) error
so the server implementation can send as many responses as it wants.

On the client side, the implementation uses a channel to read
the responses (and the for : range construct, most likely).